data_pipeline/training/requirements.txt
google-cloud-bigquery==3.13.0
matplotlib==3.10.6
pandas==2.1.4
python-dotenv==1.1.1